Don't copy if already there

+++ SharedProcessors/PkgCopyToPkgsFolder.py
@@ -35,15 +35,21 @@ class PkgCopyToPkgsFolder(Processor):
 
     def main(self):
 
+        filename = self.env.get('filename')
         pkg_path = self.env.get('pkg_path')
         RECIPE_CACHE_DIR = self.env.get('RECIPE_CACHE_DIR')
         CACHE_DIR = os.path.abspath(os.path.join(RECIPE_CACHE_DIR, os.pardir))
         Pkgs_folder = os.path.join(CACHE_DIR, 'Pkgs')
+        dest_path = os.path.join(Pkgs_folder, filename)
 
-        if not os.path.exists(Pkgs_folder):
-            os.makedirs(Pkgs_folder)
+        if os.path.exists(dest_path):
+            pass
+        else:
+            if not os.path.exists(Pkgs_folder):
+                os.makedirs(Pkgs_folder)
 
-        shutil.copy(pkg_path, Pkgs_folder)
+            shutil.copy(pkg_path, Pkgs_folder)
+            
 
 if __name__ == '__main__':
     processor = PkgCopyToPkgsFolder()
